Deliverable 2
To solve the problem, I had the user enter a start and end date in a specific format. The program would take the String and parse it into a LocalDate. from there the ChronoUnit tool was used to find the difference between the dates. then I properly formatted the outputs.